# 8276217: Harmonize StrictMath intrinsics handling