--- a/src/hotspot/cpu/x86/c1_LIRGenerator_x86.cpp
+++ b/src/hotspot/cpu/x86/c1_LIRGenerator_x86.cpp
@@ -814,11 +814,11 @@ void LIRGenerator::do_MathIntrinsic(Intrinsic* x) {
 
   switch(x->id()) {
     case vmIntrinsics::_dabs:
-      __ abs  (calc_input, calc_result, tmp);
+      __ abs(calc_input, calc_result, tmp);
       break;
     case vmIntrinsics::_dsqrt:
     case vmIntrinsics::_dsqrt_strict:
-      __ sqrt (calc_input, calc_result, LIR_OprFact::illegalOpr);
+      __ sqrt(calc_input, calc_result, LIR_OprFact::illegalOpr);
       break;
     default:
       ShouldNotReachHere();