--- a/src/hotspot/share/opto/c2compiler.cpp
+++ b/src/hotspot/share/opto/c2compiler.cpp
@@ -454,15 +454,19 @@ bool C2Compiler::is_intrinsic_supported(const methodHandle& method, bool is_virt
     if (!Matcher::match_rule_supported(Op_Whitespace)) return false;
     break;
   case vmIntrinsics::_maxF:
+  case vmIntrinsics::_maxF_strict:
     if (!Matcher::match_rule_supported(Op_MaxF)) return false;
     break;
   case vmIntrinsics::_minF:
+  case vmIntrinsics::_minF_strict:
     if (!Matcher::match_rule_supported(Op_MinF)) return false;
     break;
   case vmIntrinsics::_maxD:
+  case vmIntrinsics::_maxD_strict:
     if (!Matcher::match_rule_supported(Op_MaxD)) return false;
     break;
   case vmIntrinsics::_minD:
+  case vmIntrinsics::_minD_strict:
     if (!Matcher::match_rule_supported(Op_MinD)) return false;
     break;
   case vmIntrinsics::_writeback0:
@@ -510,6 +514,8 @@ bool C2Compiler::is_intrinsic_supported(const methodHandle& method, bool is_virt
   case vmIntrinsics::_dpow:
   case vmIntrinsics::_min:
   case vmIntrinsics::_max:
+  case vmIntrinsics::_min_strict:
+  case vmIntrinsics::_max_strict:
   case vmIntrinsics::_arraycopy:
   case vmIntrinsics::_indexOfL:
   case vmIntrinsics::_indexOfU: