@shipilev 2021-11-01
> So we have _dsqrt and_dsqrt_strict, which must be functionally identical, but we provide both names because they're part of a public API. I think this deserves an explanatory comment in the code.

Yes, no problem, added comment near intrinsic definition.