@vnkozlov 2021-11-01
Yes, I am fine with new intrinsics for them.