@shipilev 2021-11-02
> Yes, I am fine with new intrinsics for them.

All right, see new commit then.